DOC-10078 - Minor improvements and Composer setup.
It seems 4.0 can now be pulled from packagist with composer,
amending documentation to show this in the quick and full installation.

I've added a composer.json file for running our examples as well,
which makes things much easier.

All examples will need to require `autoload.php` at the beginning of the file, but that is a minor inconvenience.
Note, when running composer, a `vendor` directory with the required dependencies will be created/updated in
the root of the project.
This is how we are able to reference the `autoload.php` file.